resolving conversations
